+++ src/DnsResolver/DnsResolver.Autorest/test/Get-AzDnsResolver.Tests.ps1
@@ -44,7 +44,10 @@ Describe 'Get-AzDnsResolver' {
         Start-Sleep -Seconds 5
         Remove-AzDnsResolver -Name $dnsResolverName -ResourceGroupName $RESOURCE_GROUP_NAME
         Start-Sleep -Seconds 5
-        Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        if ($TestMode -eq "Record")
+        {
+            Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        }
     }
 
     It 'List DNS resolvers in a resource group, expected least number of DNS resolvers retrieved' {
@@ -71,6 +74,9 @@ Describe 'Get-AzDnsResolver' {
         Start-Sleep -Seconds 5
         Remove-AzDnsResolver -Name $dnsResolverName -ResourceGroupName $RESOURCE_GROUP_NAME
         Start-Sleep -Seconds 5
-        Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        if ($TestMode -eq "Record")
+        {
+            Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        }
     }
 }
