+++ src/DnsResolver/DnsResolver.Autorest/test/New-AzDnsForwardingRulesetForwardingRule.Tests.ps1
@@ -54,6 +54,9 @@ Describe 'New-AzDnsForwardingRulesetForwardingRule' {
         Start-Sleep -Seconds 5
         Remove-AzDnsResolver -Name $dnsResolverName -ResourceGroupName $RESOURCE_GROUP_NAME
         Start-Sleep -Seconds 5
-        Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        if ($TestMode -eq "Record")
+        {
+            Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        }
     }
 }
