+++ src/DnsResolver/DnsResolver.Autorest/test/New-AzDnsForwardingRulesetVirtualNetworkLink.Tests.ps1
@@ -52,6 +52,9 @@ Describe 'New-AzDnsForwardingRulesetVirtualNetworkLink' {
         Start-Sleep -Seconds 5
         Remove-AzDnsResolver -Name $dnsResolverName -ResourceGroupName $RESOURCE_GROUP_NAME
         Start-Sleep -Seconds 5
-        Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        if ($TestMode -eq "Record")
+        {
+            Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        }
     }
 }
