+++ src/DnsResolver/DnsResolver.Autorest/test/Remove-AzDnsResolver.Tests.ps1
@@ -35,6 +35,9 @@ Describe 'Remove-AzDnsResolver' {
 
         # UNDO
         Start-Sleep -Seconds 5
-        Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        if ($TestMode -eq "Record")
+        {
+            Remove-AzVirtualNetwork -Name $virtualNetworkName -ResourceGroupName $RESOURCE_GROUP_NAME -Force
+        }
     }
 }
